# inline and specialize the symbol reading