Some more decoding speedup by avoiding a function call in the hot loops, and by having a specialization in modular decoding for the no-lz77 case (which is quite common since only e1 and e8/e9 use lz77 at the moment).

Binary size goes up a bit because of this (by 100kb or so on x86), so we may want to be a bit more selective with the inlining and templating, but I think it's ok for now.

Modular decode speedup is 5-10%, also VarDCT decode gets some speedup (3% or so).

Before: (30 decode reps, single-core)
```
station.png
Encoding      kPixels    Bytes          BPP  E MP/s  D MP/s     Max norm  SSIMULACRA2   PSNR        pnorm       BPP*pnorm   QABPP   Bugs
----------------------------------------------------------------------------------------------------------------------------------------
jxl:d0:2          987  1487163   12.0458779   9.070  18.741         -nan 100.00000000  99.99   0.00000000  0.000000000000  12.046      0
jxl:d0:3          987  1416083   11.4701367   5.467   7.072         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.470      0
jxl:d0:4          987  1401819   11.3545996   1.430   5.559         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.355      0
jxl:d0:5          987  1390068   11.2594177   0.791   5.241         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.259      0
jxl:d0:6          987  1381828   11.1926744   0.525   4.762         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.193      0
jxl:d0:7          987  1370796   11.1033163   0.372   4.456         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.103      0
jxl:d1            987   268868    2.1778050   2.738  32.124   1.36113428  84.78277345  40.43   0.58483733  1.273661684303   2.964      0
jxl:d2            987   165244    1.3384606   2.971  35.156   2.41245678  75.53099932  36.49   0.99354436  1.329819958131   3.229      0
jxl:d3            987   121301    0.9825265   2.838  36.348   3.68713610  66.98580653  34.43   1.34944885  1.325869234777   3.623      0
Aggregate:        987   702916    5.6935529   1.849  11.443   2.29623621  91.02428861  71.81   0.92213164  1.309531274247   7.511      0
```

After:
```
station.png
Encoding      kPixels    Bytes          BPP  E MP/s  D MP/s     Max norm  SSIMULACRA2   PSNR        pnorm       BPP*pnorm   QABPP   Bugs
----------------------------------------------------------------------------------------------------------------------------------------
jxl:d0:2          987  1487163   12.0458779   9.257  20.856         -nan 100.00000000  99.99   0.00000000  0.000000000000  12.046      0
jxl:d0:3          987  1416083   11.4701367   5.500   7.735         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.470      0
jxl:d0:4          987  1401819   11.3545996   1.404   5.880         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.355      0
jxl:d0:5          987  1390068   11.2594177   0.793   5.513         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.259      0
jxl:d0:6          987  1381828   11.1926744   0.539   4.986         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.193      0
jxl:d0:7          987  1370796   11.1033163   0.381   4.665         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.103      0
jxl:d1            987   268868    2.1778050   2.729  33.065   1.36113428  84.78277345  40.43   0.58483733  1.273661684303   2.964      0
jxl:d2            987   165244    1.3384606   2.886  35.833   2.41245678  75.53099932  36.49   0.99354436  1.329819958131   3.229      0
jxl:d3            987   121301    0.9825265   2.821  37.419   3.68713610  66.98580653  34.43   1.34944885  1.325869234777   3.623      0
Aggregate:        987   702916    5.6935529   1.854  12.059   2.29623621  91.02428861  71.81   0.92213164  1.309531274247   7.511      0
```