@veluca93 2023-07-14
I can think of two ways:
- put `__attribute__((flatten))` (or some reasonable macro that expands to
that where supported) on hot functions, and no JXL_INLINE here.
- put `JXL_INLINE` on a renamed ReadHybridUintClusteredInline, have
ReadHybridUintClustered call that (and not be JXL_INLINE), and have "hot"
callers call ReadHybridUintClusteredInline.

I think I prefer option 2.

Also, similar changes would need to happen to `ReadHybridUint` too.

On Fri, 14 Jul 2023 at 23:10, Jon Sneyers ***@***.***> wrote:

> ***@***.**** commented on this pull request.
> ------------------------------
>
> In lib/jxl/dec_ans.h
> <https://github.com/libjxl/libjxl/pull/2665#discussion_r1264178470>:
>
> > @@ -302,53 +303,62 @@ class ANSSymbolReader {
>    }
>
>    // Takes a *clustered* idx.
> -  size_t ReadHybridUintClustered(size_t ctx, BitReader* JXL_RESTRICT br) {
> -    if (JXL_UNLIKELY(num_to_copy_ > 0)) {
> -      size_t ret = lz77_window_[(copy_pos_++) & kWindowMask];
> -      num_to_copy_--;
> -      lz77_window_[(num_decoded_++) & kWindowMask] = ret;
> -      return ret;
> +  template <bool uses_lz77>
> +  JXL_INLINE size_t ReadHybridUintClustered(size_t ctx,
>
> Most of the both comes from that. It's not ideal that it gets inlined
> everywhere now, also in the not hot parts. Is there a way to do an inlining
> macro on the call site rather than on the function definition?
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/libjxl/libjxl/pull/2665#discussion_r1264178470>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAOPAI4W5Y72BTDDBE225H3XQGYVVANCNFSM6AAAAAA2KZIE5M>
> .
> You are receiving this because your review was requested.Message ID:
> ***@***.***>
>