@jonsneyers 2023-07-14
Actually I think the only hot place that calls ReadHybridUint is the AC coeff decoding, anything else is stuff like decoding MA trees or patches or splines or other entropy-coded header stuff where we don't want inlining.
So if we just make ReadHybridUint the non-inlined one and ReadHybridUintClustered the inlined one, and change the call in the AC coeff decoding, I think that should work...

I'm currently on the train to Covilha :) but I'll try that tomorrow or so.