@jonsneyers 2023-07-15
Made some more tweaks for both speed and binary size.

Also I made a mistake in the "before" numbers above, those were made with a build that wasn't clean. Correct numbers below, where both before and after are updated.


Binary size (stripped x64 release build of libjxl.so):

Before: 2694 kb
First version of this PR: 2764 kb (+ 70 kb)
Current version of this PR: 2727 kb ( + 33 kb)

Decode speed:
Modular: 6-12% faster (more speedup for lower effort encodes)
VarDCT: 4-6% faster (more speedup for lower distance encodes)


Before:
```
station.png
Encoding      kPixels    Bytes          BPP  E MP/s  D MP/s     Max norm  SSIMULACRA2   PSNR        pnorm       BPP*pnorm   QABPP   Bugs
----------------------------------------------------------------------------------------------------------------------------------------
jxl:d0:2          987  1487163   12.0458779   8.994  18.572         -nan 100.00000000  99.99   0.00000000  0.000000000000  12.046      0
jxl:d0:3          987  1416083   11.4701367   5.414   7.070         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.470      0
jxl:d0:4          987  1401819   11.3545996   1.415   5.522         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.355      0
jxl:d0:5          987  1390068   11.2594177   0.780   5.241         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.259      0
jxl:d0:6          987  1381828   11.1926744   0.521   4.671         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.193      0
jxl:d0:7          987  1370796   11.1033163   0.365   4.466         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.103      0
jxl:d1            987   268868    2.1778050   2.759  31.539   1.36113428  84.78277345  40.43   0.58483733  1.273661684303   2.964      0
jxl:d2            987   165244    1.3384606   2.875  34.133   2.41245678  75.53099932  36.49   0.99354436  1.329819958131   3.229      0
jxl:d3            987   121301    0.9825265   2.746  35.778   3.68713610  66.98580653  34.43   1.34944885  1.325869234777   3.623      0
Aggregate:        987   702916    5.6935529   1.823  11.320   2.29623621  91.02428861  71.81   0.92213164  1.309531274247   7.511      0
```

After:
```
station.png
Encoding      kPixels    Bytes          BPP  E MP/s  D MP/s     Max norm  SSIMULACRA2   PSNR        pnorm       BPP*pnorm   QABPP   Bugs
----------------------------------------------------------------------------------------------------------------------------------------
jxl:d0:2          987  1487163   12.0458779   9.107  21.065         -nan 100.00000000  99.99   0.00000000  0.000000000000  12.046      0
jxl:d0:3          987  1416083   11.4701367   5.464   7.656         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.470      0
jxl:d0:4          987  1401819   11.3545996   1.411   5.873         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.355      0
jxl:d0:5          987  1390068   11.2594177   0.786   5.519         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.259      0
jxl:d0:6          987  1381828   11.1926744   0.525   5.029         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.193      0
jxl:d0:7          987  1370796   11.1033163   0.371   4.726         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.103      0
jxl:d1            987   268868    2.1778050   2.767  33.473   1.36113428  84.78277345  40.43   0.58483733  1.273661684303   2.964      0
jxl:d2            987   165244    1.3384606   2.943  35.891   2.41245678  75.53099932  36.49   0.99354436  1.329819958131   3.229      0
jxl:d3            987   121301    0.9825265   2.814  37.344   3.68713610  66.98580653  34.43   1.34944885  1.325869234777   3.623      0
Aggregate:        987   702916    5.6935529   1.844  12.104   2.29623621  91.02428861  71.81   0.92213164  1.309531274247   7.511      0
```